# openapi/cross-platform.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GamePedia Cross-Platform Gate API",
    "version": "2.0.0",
    "description": "Machine-readable subset for the GamePedia 2.0 iOS/Android contract gate. This is not the complete backend API. Dates are RFC 3339 UTC date-time strings; image URLs are absolute HTTP(S) URLs or null."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Get the authenticated account",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The current account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` (missing/invalid bearer token) or `TOKEN_EXPIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`ACCOUNT_INACTIVE` or `ACCOUNT_SUSPENDED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`ACCOUNT_NOT_FOUND`: the account behind the token no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentUser",
        "summary": "Delete the authenticated account",
        "description": "Transactionally deletes the account and all of its refresh sessions (plus dependent rows via ON DELETE CASCADE). Afterwards no access or refresh token of the user can authenticate.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Account deleted; all sessions are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthAccountDeletedSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` or `TOKEN_EXPIRED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`ACCOUNT_INACTIVE` or `ACCOUNT_SUSPENDED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`ACCOUNT_NOT_FOUND`: already deleted (repeat calls are rejected, not idempotent-success).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "summary": "Get current user profile",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ProfileSuccess"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "summary": "Update current user profile",
        "responses": {
          "200": {
            "$ref": "#/components/responses/ProfileSuccess"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/users/me/privacy": {
      "get": {
        "summary": "Get privacy settings",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PrivacySuccess"
          }
        }
      },
      "patch": {
        "summary": "Update privacy settings",
        "requestBody": {
          "$ref": "#/components/requestBodies/PrivacyUpdate"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PrivacySuccess"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/users/me/privacy-settings": {
      "get": {
        "deprecated": true,
        "summary": "iOS compatibility alias for /users/me/privacy",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PrivacySuccess"
          }
        }
      },
      "patch": {
        "deprecated": true,
        "summary": "iOS compatibility alias for /users/me/privacy",
        "requestBody": {
          "$ref": "#/components/requestBodies/PrivacyUpdate"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PrivacySuccess"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/users/me/recently-played": {
      "get": {
        "summary": "Get current user's recently played games",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RecentPlaySuccess"
          }
        }
      }
    },
    "/users/me/recent-plays": {
      "get": {
        "deprecated": true,
        "summary": "iOS compatibility alias for /users/me/recently-played",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RecentPlaySuccess"
          }
        }
      }
    },
...
    "/users/me/friends/steam/import": {
      "post": {
        "summary": "Refresh Steam friend discovery without creating friendships",
        "description": "Idempotent discovery operation. It never auto-creates GamePedia friendship rows.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SteamFriendImportSuccess"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/users/{userId}/friend-recommendations": {
      "get": {
        "summary": "Recommend games using one accepted friend's visible signals",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RecommendationSuccess"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/users/me/push-token": {
      "put": {
        "summary": "Register or transfer a push token",
        "requestBody": {
          "$ref": "#/components/requestBodies/PushToken"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PushRegistrationSuccess"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "summary": "Compatibility registration method",
        "requestBody": {
          "$ref": "#/components/requestBodies/PushToken"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PushRegistrationSuccess"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Deactivate a push token",
        "description": "Accepts deviceId/token in query for deployed iOS compatibility or in a JSON body; conflicting duplicate inputs are rejected.",
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "maxLength": 200
            }
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "minLength": 20,
              "maxLength": 4096
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PushDeleteSuccess"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable error code. Search-related codes: INVALID_SEARCH_QUERY, INVALID_GAMES_LIMIT, VALIDATION_ERROR, IGDB_RATE_LIMITED, IGDB_UPSTREAM_ERROR, TWITCH_AUTH_UNAVAILABLE, IGDB_NOT_CONFIGURED, INTERNAL_SERVER_ERROR, NOT_FOUND. Auth-related codes: UNAUTHORIZED, TOKEN_EXPIRED, TOKEN_REVOKED, INVALID_CREDENTIALS, ACCOUNT_NOT_FOUND, ACCOUNT_INACTIVE, ACCOUNT_SUSPENDED, EMAIL_ALREADY_IN_USE, NICKNAME_ALREADY_EXISTS, PASSWORD_RESET_TOKEN_INVALID, PASSWORD_RESET_TOKEN_USED, PASSWORD_RESET_TOKEN_EXPIRED, APPLE_IDENTITY_TOKEN_REQUIRED, APPLE_EMAIL_REQUIRED, APPLE_EMAIL_NOT_VERIFIED, APPLE_AUTH_INVALID_TOKEN, APPLE_IDENTITY_TOKEN_EXPIRED, APPLE_AUTH_UNAVAILABLE, APPLE_AUTH_NOT_CONFIGURED, GOOGLE_ID_TOKEN_REQUIRED, GOOGLE_EMAIL_REQUIRED, GOOGLE_EMAIL_NOT_VERIFIED, GOOGLE_AUTH_INVALID_TOKEN, GOOGLE_ID_TOKEN_EXPIRED, GOOGLE_AUTH_UNAVAILABLE, GOOGLE_AUTH_NOT_CONFIGURED, GOOGLE_ACCOUNT_LINK_CONFLICT, SOCIAL_ACCOUNT_CONFLICT, CONFLICT."
              },
              "message": {
                "type": "string",
                "description": "Human-readable message. Sanitized server-side; may be localized. Clients must branch on `code`, never on `message`."
              },
              "details": {
                "description": "Optional structured details. For validation failures this is an array of {field, message}.",
                "type": [
                  "array",
                  "object",
                  "null"
                ]
              }
            }
          }
        }
      },
...
      "AuthUser": {
        "type": "object",
        "description": "Public representation of the authenticated account. Never contains password hashes or token material.",
        "required": [
          "id",
          "email",
          "nickname",
          "profileImageUrl",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "nickname": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30
          },
          "profileImageUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Absolute URL when set; null when the account has no profile image."
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 UTC timestamp."
          }
        }
      },
...
      "AuthMeSuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "user"
            ],
            "properties": {
              "user": {
                "$ref": "#/components/schemas/AuthUser"
              }
            }
          }
        }
      },
...
      "AuthAccountDeletedSuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "deleted",
              "deletedAt"
            ],
            "properties": {
              "deleted": {
                "const": true
              },
              "deletedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
...
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {}
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": [
                  "object",
                  "array",
                  "null"
                ]
              }
            }
          }
        }
      },
...
      "ProfileData": {
        "type": "object",
        "required": [
          "user"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "friendCount": {
            "type": "integer",
            "minimum": 0
          },
          "likeCount": {
            "type": "integer",
            "minimum": 0
          },
          "reviewCount": {
            "type": "integer",
            "minimum": 0
          },
          "recentlyPlayed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentGame"
            }
          },
          "hasMoreRecentPlayed": {
            "type": "boolean"
          }
        }
      },
      "PrivacyUpdate": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "showFriendsList": {
            "type": "boolean"
          },
          "showRecentlyPlayed": {
            "type": "boolean"
          },
          "showLikedGames": {
            "type": "boolean"
          },
          "showReviews": {
            "type": "boolean"
          },
          "isFriendsListPublic": {
            "type": "boolean",
            "deprecated": true
          },
          "isRecentPlayPublic": {
            "type": "boolean",
            "deprecated": true
          },
          "isLikedGamesPublic": {
            "type": "boolean",
            "deprecated": true
          },
          "isReviewsPublic": {
            "type": "boolean",
            "deprecated": true
          }
        }
      },
      "PrivacySettings": {
        "type": "object",
        "required": [
          "showFriendsList",
          "showRecentlyPlayed",
          "showLikedGames",
          "showReviews",
          "isFriendsListPublic",
          "isRecentPlayPublic",
          "isLikedGamesPublic",
          "isReviewsPublic"
        ],
        "properties": {
          "privacy": {
            "type": "object"
          },
          "showFriendsList": {
            "type": "boolean"
          },
          "showRecentlyPlayed": {
            "type": "boolean"
          },
          "showLikedGames": {
            "type": "boolean"
          },
          "showReviews": {
            "type": "boolean"
          },
          "isFriendsListPublic": {
            "type": "boolean"
          },
          "isRecentPlayPublic": {
            "type": "boolean"
          },
          "isLikedGamesPublic": {
            "type": "boolean"
          },
          "isReviewsPublic": {
            "type": "boolean"
          },
          "steamFriendsFeatureAvailable": {
            "type": "boolean"
          }
        }
      },
...
      "RecentPlayData": {
        "type": "object",
        "required": [
          "games",
          "recentGames",
          "recentlyPlayed",
          "recentPlayedPreview",
          "hasMoreRecentPlayed"
        ],
        "properties": {
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentGame"
            }
          },
          "recentGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentGame"
            }
          },
          "recentlyPlayed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentGame"
            }
          },
          "recentPlayedPreview": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentGame"
            }
          },
          "hasMoreRecentPlayed": {
            "type": "boolean"
          }
        }
      },
      "FriendRecommendation": {
        "type": "object",
        "required": [
          "game"
        ],
        "properties": {
          "game": {
            "type": "object"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PushTokenDeleteRequest": {
        "type": "object",
        "anyOf": [
          {
            "required": [
              "deviceId"
            ]
          },
          {
            "required": [
              "token"
            ]
          }
        ],
        "properties": {
          "deviceId": {
            "type": "string",
            "maxLength": 200
          },
          "token": {
            "type": "string",
            "minLength": 20,
            "maxLength": 4096,
            "writeOnly": true
          }
        }
      },
      "PushRegistrationData": {
        "type": "object",
        "required": [
          "registered"
        ],
        "properties": {
          "registered": {
            "const": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT access token issued by the auth endpoints. Sent as `Authorization: Bearer <accessToken>`."
      }
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50,
          "default": 5
        }
      },
      "UserId": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "requestBodies": {
      "PrivacyUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PrivacyUpdate"
            }
          }
        }
      },
      "PushToken": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PushTokenRequest"
            }
          }
        }
      }
    },
    "responses": {
...
      "Error": {
        "description": "Shared error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "ProfileSuccess": {
        "description": "Current user profile",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProfileData"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "SteamFriendImportSuccess": {
        "description": "Steam friend discovery result",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SteamFriendImportData"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "PushRegistrationSuccess": {
        "description": "Push token registered",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PushRegistrationData"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "PushDeleteSuccess": {
        "description": "Push token deactivated",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PushDeleteData"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "PrivacySuccess": {
        "description": "Privacy settings",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PrivacySettings"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "RecentPlaySuccess": {
        "description": "Recently played page",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RecentPlayData"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "RecommendationSuccess": {
        "description": "Recommendations",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "recommendations"
                      ],
                      "properties": {
                        "recommendations": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FriendRecommendation"
                          }
                        }
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}